Rename Arithmetic.cpp to 03/arithmetic.cpp
